Rename rpc_server module
